LibWeb: Split FrequencyStyleValue out of StyleValue.{h,cpp}
