+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -661,30 +661,6 @@ private:
     NonnullOwnPtr<CalcSum> m_expression;
 };
 
-class FrequencyStyleValue : public StyleValueWithDefaultOperators<FrequencyStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<FrequencyStyleValue> create(Frequency frequency)
-    {
-        return adopt_ref(*new FrequencyStyleValue(move(frequency)));
-    }
-    virtual ~FrequencyStyleValue() override { }
-
-    Frequency const& frequency() const { return m_frequency; }
-
-    virtual ErrorOr<String> to_string() const override { return m_frequency.to_string(); }
-
-    bool properties_equal(FrequencyStyleValue const& other) const { return m_frequency == other.m_frequency; };
-
-private:
-    explicit FrequencyStyleValue(Frequency frequency)
-        : StyleValueWithDefaultOperators(Type::Frequency)
-        , m_frequency(move(frequency))
-    {
-    }
-
-    Frequency m_frequency;
-};
-
 class GridTemplateAreaStyleValue final : public StyleValueWithDefaultOperators<GridTemplateAreaStyleValue> {
 public:
     static ValueComparingNonnullRefPtr<GridTemplateAreaStyleValue> create(Vector<Vector<String>> grid_template_area);
